Refactored Displayable and implemented it
I realized by combining ToPoints/ToPixels/ToScaled, the documentation
and usage becomes much clearer. I originally thought it was going to
require too many impls, but in the end, I realized it was the same
number of function implementations in the end.

Now, all types that have a unit implement Displayable.