Refactor Lp to use a smaller scale
I realized with all the numbers being used, this was larger than it
needed to be since related math operations could be adjusted to using
different ratios.

In the process of fixing this and documenting Lp better, I noticed
fraction * i32/u32 had flipped the numerator and denominator, likely a
copy/paste bug from an early div implementation.